perf(firmware): no 64-bit integer helpers in the overview maths (rc3)
On the radar, rc2 returned only the l3overview echo: it computes before its
first byte, and the Pi's reader counted its 4 s stall timer from the echo.
Compiling rc2's core with the TI toolchain showed per-value calls to
__aeabi_lmul and __aeabi_l2d -- the R4F has no 64-bit multiply or
int64-to-double instruction.

* reduced_overview.c: every intermediate is an exact integer below 2^53,
  so it is carried in doubles -- bit-identical results, no lmul/l2d. The
  median's key is kept as two 32-bit halves so the per-value work needs no
  64-bit shift helper (those now run once per pass). Still byte-identical
  to reduced.py on all 80 saved captures; a uniform-capture test drives the
  median through all 64 key bits.
* driver: for l3overview/l3strip the stall timer starts only once the
  response has begun, and a missing response raises instead of returning
  the echo. l3dump keeps its contract (tested).
* hardware test: --overview-timeout (default 30 s) so a slow build still
  reports its timing.

Candidate: firmware/candidates/l3_dump_reduced_transfer_rc3.bin
(SHA-256 a1669b87...), replacing rc2.

+++ firmware/iwr6843/reduced_overview.h
@@ -71,11 +71,15 @@ typedef struct {
     uint8_t count;
 } ro_window_t;
 
-/* Working memory for the overview (~104 KB; keep it off the stack). */
+/* Working memory for the overview (~104 KB; keep it off the stack).
+ *
+ * The R4F has no 64-bit integer multiply-to-double path, so exact integer
+ * quantities (all below 2^53 for every supported capture) are held in
+ * doubles, where they stay exact. */
 typedef struct {
     /* Window scope: scaled code sums T and loop counts n per absolute bin. */
-    int64_t  total_re[2][RO_MAX_RX][RO_BIN_SPACE];
-    int64_t  total_im[2][RO_MAX_RX][RO_BIN_SPACE];
+    double   total_re[2][RO_MAX_RX][RO_BIN_SPACE];
+    double   total_im[2][RO_MAX_RX][RO_BIN_SPACE];
     uint32_t total_count[RO_BIN_SPACE];
     double   inv_nn[RO_BIN_SPACE];               /* 1 / n^2 */
     /* One frame's vertical-pair codes [pair][loop][rx][bin] and loop sums. */
